Remove category menu

+++ EcommerceApp.Core/Models/Products/ProductInfo.cs
@@ -15,7 +15,7 @@
 
         public IEnumerable<ProductStock<T>> ProductStocks { get; set; }
         public string Brand { get; set; } = null!;
-        public double AverageRating => Reviews.Count() > 0 ? Reviews.Sum(r => r.StarEvaluation) / Reviews.Count() : 0;
+        public double AverageRating => Reviews.Count() > 0 ? (double)Reviews.Sum(r => r.StarEvaluation) / Reviews.Count() : 0;
         public int TotalReviewsCount => Reviews.Count();
         public string Gender { get; set; } = null!;
         public IEnumerable<ProductModel> RelatedProducts { get; set; }
